Migrate localization clip selection into LocalizationClipStrategy
Move all localization logic (table/entry validation, asset DB lookup,
locale matching, wrapper creation) from AudioEntity.PickLocalizationClip
into LocalizationClipStrategy.SelectClip. The entity now injects table,
entry, and name into the strategy via Inject(), then delegates through
the shared EnsureClipSelectionStrategy/SelectClip path like all other
play modes.
